requirements.txt
absl-py==2.0.0
accelerate==1.6.0
addict==2.4.0
aenum==3.1.15
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.1
albumentations==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.1.2
annotated-types==0.6.0
ansicon==1.89.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
av==14.3.0
babel==2.17.0
basicsr==1.4.2
beautifulsoup4==4.12.3
blend_modes==2.2.0
blendmodes==2023
blessed==1.21.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.2.0
codeformer==0.0.11
colorama==0.4.6
coloredlogs==15.0.1
comfy==0.0.1
comfyui_frontend_package==1.19.9
comfyui_workflow_templates==0.1.14
comtypes==1.4.10
contourpy==1.2.0
crcmod==1.7
cryptography==41.0.5
csvw==3.5.1
customtkinter==5.2.2
cv2_enumerate_cameras==1.1.18.3
cycler==0.12.1
Cython==3.0.8
darkdetect==0.8.0
datasets==2.13.0
decorator==4.4.2
deepdiff==8.5.0
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.33.1
dill==0.3.6
dlinfo==2.0.0
dnspython==2.7.0
docstring_parser==0.16
dynamicprompts==0.31.0
easydict==1.11
editor==1.6.6
einops==0.7.0
evalidate==2.0.5
exceptiongroup==1.2.0
facexlib==0.3.0
fairscale==0.4.13
fastapi==0.115.12
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.14.0
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.3
future==0.18.3
gast==0.4.0
gdown==5.0.0
gfpgan==1.3.8
gguf==0.9.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
gradio==5.29.0
gradio_client==1.10.0
groovy==0.1.2
grpcio==1.59.3
h11==0.16.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.6
ifaddr==0.2.0
imageio==2.33.0
imageio-ffmpeg==0.6.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
inquirer==3.4.0
insightface==0.7.3
intel-openmp==2021.4.0
isodate==0.7.2
jax==0.4.23
jaxlib==0.4.23
Jinja2==3.1.2
jinxed==1.3.0
jmespath==0.10.0
joblib==1.3.2
jsonmerge==1.9.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
keras==2.15.0
keras-nightly==3.4.1.dev2024071603
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kornia==0.8.1
kornia_rs==0.1.9
language-tags==1.2.0
lark==1.1.8
latexcodec==3.0.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.0
lightning-utilities==0.10.0
llvmlite==0.41.1
lmdb==1.4.1
lpips==0.1.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matminer==0.9.3
matplotlib==3.7.3
mdurl==0.1.2
mediapipe==0.10.21
meshio==5.3.5
mkl==2021.4.0
ml-dtypes==0.2.0
mmcv-full==1.7.2
mmengine==0.10.4
model-index==0.1.11
modelscope==1.9.5
monty==2025.3.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.4
multiprocess==0.70.14
namex==0.0.8
narwhals==1.38.2
networkx==3.2.1
ninja==1.11.1.1
norbert==0.2.1
numba==0.58.1
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.0
omegaconf==2.3.0
onnx==1.14.0
onnxruntime-gpu==1.22.0
open-clip-torch==2.23.0
opencv-contrib-python==4.10.0.82
opencv-contrib-python-headless==4.11.0.86
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
opennsfw2==0.10.2
openxlab==0.1.1
opt-einsum==3.3.0
optree==0.12.1
ordered-set==4.1.0
orderly-set==5.4.1
orjson==3.9.10
oss2==2.17.0
packaging==24.1
palettable==3.3.3
pandas==1.5.0
phonemizer==3.3.0
piexif==1.1.3
pillow==10.2.0
platformdirs==4.0.0
plotly==6.0.1
pooch==1.8.2
praat-parselmouth==0.4.5
prettytable==3.9.0
proglog==0.1.11
propcache==0.3.1
protobuf==4.23.2
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybtex==0.24.0
pycocotools==2.0.8
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
pyfbx==0.0.8
pygame==2.5.2
Pygments==2.18.0
pygrabber==0.2
pymatgen==2024.5.1
pymongo==4.12.1
PyMuPDF==1.24.13
pyparsing==3.1.1
pyreadline3==3.4.1
PySocks==1.7.1
pytest-runner==6.0.1
python-dateutil==2.8.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.1.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
qudida==0.0.4
rdflib==7.1.4
readchar==4.2.1
realesrgan==0.3.0
referencing==0.31.0
regex==2023.10.3
requests==2.32.3
requests-oauthlib==1.3.1
resampy==0.4.3
resize-right==0.0.2
rfc3986==1.5.0
rich==13.4.2
rpds-py==0.13.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.9
runs==1.2.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.2
segment-anything==1.0
segments==2.3.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.2
simpleeval==0.9.13
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.5
soxr==0.5.0.post1
spandrel==0.4.1
spglib==2.6.0
spleeter==2.4.0
starlette==0.46.2
supervision==0.25.1
surrealist==1.1.2
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.18.0a20240716
tbb==2021.11.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf-nightly==2.18.0.dev20240713
tf_keras-nightly==2.18.0.dev2024071609
tf_nightly_intel==2.18.0.dev20240713
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.12
tk==0.1.0
tkinterdnd2==0.3.0
tokenizers==0.21.1
tomesd==0.1.3
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2+cu118
torchdiffeq==0.2.3
torchmetrics==1.2.0
torchsde==0.2.6
torchvision==0.15.2+cu118
tqdm==4.67.1
trampoline==0.1.2
transformers==4.51.3
typeguard==4.4.2
typer==0.15.3
typing_extensions==4.13.2
tyro==0.9.19
tzdata==2023.3
ultralytics==8.0.176
ultralytics-thop==2.0.0
uncertainties==3.2.3
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.24.0.post1
vtracer==0.6.11
wcwidth==0.2.12
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.1
wget==3.2
whisper==1.1.10
wrapt==1.14.1
xmod==1.8.1
xxhash==3.4.1
yapf==0.40.2
yarl==1.20.0
zeroconf==0.132.2
zipp==3.17.0
